backend/requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# API & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# LangGraph & Agents
langgraph>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0

# Composio
composio-openai>=0.1.0
composio-langchain>=0.1.0

# Vector Store
pinecone-client>=3.0.0
chromadb>=0.4.0

# LLMs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Caching
redis>=5.0.0

# Utilities
python-dateutil>=2.8.0
aiofiles>=23.2.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
